refactoring code to upload image to separate folder
