+++ frontend/pantry/src/components/Upload/styles.ts
